changed app and currency
Доделала логику расчета валюты, перенесла расчеты в App.js
